test: defined a shared mock for chrome
