+++ packages/shared/package.json
@@ -7,7 +7,7 @@
     "tsc": "tsc",
     "build": "tsc -b",
     "lint": "eslint ./src",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "license": "AGPL-3.0",
   "dependencies": {
