+++ platforms/tktrex/extension/__tests__/app.spec.ts
@@ -1,22 +1,17 @@
-import axiosMock from '@shared/test/__mocks__/axios.mock';
-import { boot } from '@shared/extension/app';
-import { handleServerLookup } from '@shared/extension/chrome/background/account';
+import { HandshakeActiveResponseArb } from '@shared/arbitraries/HandshakeResponse.arb';
+import { boot, BootOpts } from '@shared/extension/app';
 import { load } from '@shared/extension/chrome/background/index';
-import { handleSyncMessage } from '@shared/extension/chrome/background/sync';
+import { getChromeMock } from '@shared/extension/__mocks__/chrome';
+import { fc } from '@shared/test';
+import axiosMock from '@shared/test/__mocks__/axios.mock';
+import { sleep } from '@shared/utils/promise.utils';
 import { tiktokDomainRegExp } from '@tktrex/parser/constant';
 import * as fs from 'fs';
-import { chrome } from 'jest-chrome';
 import * as path from 'path';
 import * as app from '../src/app/app';
 import * as handlers from '../src/app/handlers';
 import api, { getHeadersForDataDonation } from '../src/background/api';
 import tkHub from '../src/handlers/hub';
-import { tkLog } from '../src/logger';
-import { sleep } from '@shared/utils/promise.utils';
-import { HandshakeActiveResponseArb } from '@shared/arbitraries/HandshakeResponse.arb';
-import { fc } from '@shared/test';
-
-const chromeListener = jest.fn();
 
 const videoMatcher = app.tkHandlers.video;
 const searchMatcher = app.searchHandler;
@@ -28,8 +23,8 @@ const handleSearchSpy = jest.spyOn(searchMatcher, 'handle');
 
 const researchTag = 'fake-tag';
 const keys = {
-  publicKey: process.env.PUBLIC_KEY,
-  secretKey: process.env.SECRET_KEY,
+  publicKey: process.env.PUBLIC_KEY as any,
+  secretKey: process.env.SECRET_KEY as any,
 };
 
 const getConfig = () => ({
@@ -39,11 +34,12 @@ const getConfig = () => ({
   href: tkURL,
   researchTag,
   execount: 1,
+  experimentId: undefined,
   testTime: new Date().toISOString(),
 });
 
 const tkURL = 'https://tiktok.com/foryou';
-const bootOptions = {
+const bootOptions: BootOpts = {
   payload: {
     config: { ...keys, researchTag },
     href: window.location.href,
@@ -61,77 +57,15 @@ const bootOptions = {
   onAuthenticated: app.tkTrexActions,
 };
 
-chromeListener.mockImplementation((request, sender, sendResponse) => {
-  tkLog.info('on listener %O', { request, sender });
-  sendResponse({ type: 'Success', response: null });
-  return true;
-});
-
-chrome.runtime.sendMessage
-  // mock first 'chromeConfig' message handler
-  // .mockImplementationOnce((msg, cb: any) => {
-  //   return cb({ ux: true, active: true });
-  // })
-  // mock 'LocalLookup' message handler
-  .mockImplementation((msg: any, cb: any) => {
-    // mock 'SettingsLookup' message handler
-    if (msg.type === 'SettingsLookup') {
-      const settings = getConfig();
-      dbMap.local = settings as any;
-      return cb(settings);
-      // mock 'LocalLookup' message handler
-    } else if (msg.type === 'LocalLookup') {
-      return cb(getConfig());
-    } else if (msg.type === 'ServerLookup') {
-      return handleServerLookup(backgroundOpts)(msg.payload, cb);
-    } else if (msg.type === 'sync') {
-      return handleSyncMessage(backgroundOpts)(msg, null, cb);
-    }
-    return cb({
-      type: 'Error',
-      error: new Error(`Message type ${msg.type} not handled`),
-    });
-  });
-
-let dbMap = {
-  local: undefined,
-};
-
-chrome.storage.local.get.mockImplementation((key: any, cb: any) => {
-  app.appLog.info('Get Storage key (%s) %O', key, dbMap);
-  return cb(dbMap);
-});
-
-chrome.storage.local.set.mockImplementation((obj: any, cb: any) => {
-  app.appLog.info('Set Storage key %O', obj);
-  dbMap = {
-    ...dbMap,
-    ...obj,
-  };
-  return cb();
-});
-
-// mock connect port for 'ConfigUpdate'
-const portMock = {
-  name: 'ConfigUpdate',
-  onMessage: {
-    addListener: jest.fn(),
-    hasListener: jest.fn(),
-    getRules: jest.fn(),
-    removeRules: jest.fn(),
-  },
-  postMessage: jest.fn(),
-  disconnect: jest.fn(),
-  onDisconnect: jest.fn(),
-};
-
-chrome.runtime.connect.mockReturnValue(portMock as any);
-chrome.runtime.onMessage.addListener(chromeListener);
-
 const backgroundOpts = {
   api: api.API,
   getHeadersForDataDonation,
 };
+
+const chrome = getChromeMock({
+  getConfig,
+  backgroundOpts,
+});
 
 describe('TK App', () => {
   jest.setTimeout(20 * 1000);
